+++ src/main/java/com/testehan/finana/model/valuation/growth/IncomeStatementYear.java
@@ -2,13 +2,15 @@ package com.testehan.finana.model.valuation.growth;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class IncomeStatementYear {
 
     private int fiscalYear;
 
-    private double revenue;
-    private double operatingIncome;
-    private double pretaxIncome;
-    private double netIncome;
+    private BigDecimal revenue;
+    private BigDecimal operatingIncome;
+    private BigDecimal pretaxIncome;
+    private BigDecimal netIncome;
 }
